# frontend/openapi/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "SampleForge",
		"description": "\nMachine-readable sample generation with **hard, verified constraints**.\n\nSubmit a `GenerationSpec` and get back a job id; poll it for rendered files and\na per-constraint pass/fail map. Every spectral and temporal constraint is\nenforced during rendering and measured afterwards, so `constraints_met` is\nevidence rather than intent.\n\n* `POST /v1/generate` \u2014 queue one job\n* `GET /v1/jobs/{job_id}` \u2014 poll it\n* `POST /v1/batch`, `GET /v1/batches/{batch_id}` \u2014 many at once\n* `/v1/recipes` \u2014 save and reuse specs\n* `GET /v1/health` \u2014 liveness (unauthenticated)\n\nAuthenticated endpoints take an `X-API-Key` header. Recipe reads are open;\neverything else requires a key.\n",
		"version": "0.1.0"
	},
	"paths": {
...
		"/v1/jobs/{job_id}": {
			"get": {
				"tags": ["jobs"],
				"summary": "Poll a generation job",
				"description": "Returns the job's current status. While `queued` or `processing` the `files` list is empty. On `complete` or `complete_with_warnings` it carries every rendered variant's URL plus `constraints_met`, the per-constraint pass/fail map.",
				"operationId": "get_job_v1_jobs__job_id__get",
				"parameters": [
					{
						"name": "job_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Job Id"
						}
					},
					{
						"name": "X-API-Key",
						"in": "header",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "API key identifying the caller.",
							"auto_error": false,
							"title": "X-Api-Key"
						},
						"description": "API key identifying the caller."
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenerationResponse"
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid API key."
					},
					"404": {
						"description": "No such job."
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
...
		"/v1/recipes": {
			"post": {
				"tags": ["recipes"],
				"summary": "Save a recipe",
				"description": "Stores a `GenerationSpec` under a unique name. The spec is validated on the way in, so a recipe can never be saved in a state the generator would reject.",
				"operationId": "create_recipe_v1_recipes_post",
				"parameters": [
					{
						"name": "X-API-Key",
						"in": "header",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "API key identifying the caller.",
							"auto_error": false,
							"title": "X-Api-Key"
						},
						"description": "API key identifying the caller."
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RecipeCreate"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecipeRead"
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid API key."
					},
					"409": {
						"description": "A recipe with that name exists."
					},
					"422": {
						"description": "The spec is invalid."
					}
				}
			},
			"get": {
				"tags": ["recipes"],
				"summary": "List recipes",
				"description": "Newest first, paginated. `limit` is capped at 100 so a single request cannot pull the whole table.",
				"operationId": "list_recipes_v1_recipes_get",
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"maximum": 100,
							"minimum": 1,
							"description": "Page size.",
							"default": 20,
							"title": "Limit"
						},
						"description": "Page size."
					},
					{
						"name": "offset",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 0,
							"description": "Recipes to skip.",
							"default": 0,
							"title": "Offset"
						},
						"description": "Recipes to skip."
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecipeList"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/v1/recipes/{recipe_id}": {
			"get": {
				"tags": ["recipes"],
				"summary": "Fetch a recipe",
				"description": "Return one saved recipe.\n\nArgs:\n    request: Incoming request.\n    recipe_id: The recipe to fetch.\n    session: Async database session.\n\nReturns:\n    The recipe.\n\nRaises:\n    HTTPException: 404 if it does not exist.",
				"operationId": "get_recipe_v1_recipes__recipe_id__get",
				"parameters": [
					{
						"name": "recipe_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Recipe Id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecipeRead"
								}
							}
						}
					},
					"404": {
						"description": "No such recipe."
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["recipes"],
				"summary": "Update a recipe",
				"description": "Partial update: send only the fields to change. Renaming to an existing name is a 409 rather than a silent overwrite.",
				"operationId": "update_recipe_v1_recipes__recipe_id__put",
				"parameters": [
					{
						"name": "recipe_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Recipe Id"
						}
					},
					{
						"name": "X-API-Key",
						"in": "header",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "API key identifying the caller.",
							"auto_error": false,
							"title": "X-Api-Key"
						},
						"description": "API key identifying the caller."
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RecipeUpdate"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecipeRead"
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid API key."
					},
					"404": {
						"description": "No such recipe."
					},
					"409": {
						"description": "That name is already taken."
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["recipes"],
				"summary": "Delete a recipe",
				"description": "Removes the recipe. Jobs already generated from it are untouched \u2014 the spec is denormalised onto each job row, so history stays reproducible.",
				"operationId": "delete_recipe_v1_recipes__recipe_id__delete",
				"parameters": [
					{
						"name": "recipe_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Recipe Id"
						}
					},
					{
						"name": "X-API-Key",
						"in": "header",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "API key identifying the caller.",
							"auto_error": false,
							"title": "X-Api-Key"
						},
						"description": "API key identifying the caller."
					}
				],
				"responses": {
					"204": {
						"description": "Successful Response"
					},
					"401": {
						"description": "Missing or invalid API key."
					},
					"404": {
						"description": "No such recipe."
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/v1/recipes/{recipe_id}/run": {
			"post": {
				"tags": ["recipes"],
				"summary": "Generate from a saved recipe",
				"description": "Queues a job using the recipe's stored spec. The job records the recipe it came from, so generated files can be traced back to the exact spec that produced them.",
				"operationId": "run_recipe_v1_recipes__recipe_id__run_post",
				"parameters": [
					{
						"name": "recipe_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Recipe Id"
						}
					},
					{
						"name": "X-API-Key",
						"in": "header",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "API key identifying the caller.",
							"auto_error": false,
							"title": "X-Api-Key"
						},
						"description": "API key identifying the caller."
					}
				],
				"responses": {
					"202": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecipeRunResponse"
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid API key."
					},
					"404": {
						"description": "No such recipe."
					},
					"503": {
						"description": "The queue is unreachable."
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
...
			"GeneratedFile": {
				"properties": {
					"url": {
						"type": "string",
						"minLength": 1,
						"title": "Url",
						"description": "Download URL for the rendered file (typically a signed URL)."
					},
					"duration_ms": {
						"anyOf": [
							{
								"type": "integer",
								"exclusiveMinimum": 0.0
							},
							{
								"type": "null"
							}
						],
						"title": "Duration Ms",
						"description": "Measured length of the rendered file, when analysed."
					},
					"peak_db": {
						"anyOf": [
							{
								"type": "number",
								"maximum": 0.0
							},
							{
								"type": "null"
							}
						],
						"title": "Peak Db",
						"description": "Measured peak level in dBFS, when analysed."
					},
					"spectral_analysis": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/SpectralAnalysis"
							},
							{
								"type": "null"
							}
						],
						"description": "Post-generation spectral and temporal measurements, when analysed."
					}
				},
				"additionalProperties": false,
				"type": "object",
				"required": ["url"],
				"title": "GeneratedFile",
				"description": "A single rendered variant produced by a generation job.\n\nOnly :attr:`url` is guaranteed: a job row records where each variant landed,\nand the measurements are filled in when the analysis pass runs over the\nstored file. They are optional rather than zeroed so a consumer can tell\n\"not measured\" from \"measured as zero\" \u2014 a peak of 0 dBFS is a clipped\nmaster, not a missing value."
			},
			"GenerationResponse": {
				"properties": {
					"job_id": {
						"type": "string",
						"title": "Job Id",
						"description": "Identifier of the generation job."
					},
					"status": {
						"type": "string",
						"enum": ["queued", "processing", "complete", "complete_with_warnings", "failed"],
						"title": "Status",
						"description": "Current or final job status."
					},
					"files": {
						"items": {
							"$ref": "#/components/schemas/GeneratedFile"
						},
						"type": "array",
						"title": "Files",
						"description": "Rendered variants, one per successful sample in the batch."
					},
					"constraints_met": {
						"additionalProperties": {
							"type": "boolean"
						},
						"type": "object",
						"title": "Constraints Met",
						"description": "Per-constraint pass/fail map, keyed by constraint name (e.g. {'fundamental_hz': true, 'spectral_ceiling_hz': true}). Built by GenerationSpec.evaluate_constraints, so when a job completes normally it carries a verdict on every hard constraint."
					},
					"spec": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/GenerationSpec"
							},
							{
								"type": "null"
							}
						],
						"description": "The spec this job was submitted with, echoed back so a client can rebuild overlays and reports without having kept its own copy. None only when the stored spec no longer validates."
					}
				},
				"additionalProperties": false,
				"type": "object",
				"required": ["job_id", "status"],
				"title": "GenerationResponse",
				"description": "Result of a generation job, returned by the API and the worker.\n\nOne entry is present in :attr:`files` per generated variant, so\n``len(files)`` should equal the spec's ``batch_size`` on success."
			},
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
...
			"RecipeCreate": {
				"properties": {
					"name": {
						"type": "string",
						"maxLength": 128,
						"minLength": 1,
						"title": "Name",
						"description": "Unique recipe name."
					},
					"description": {
						"anyOf": [
							{
								"type": "string",
								"maxLength": 512
							},
							{
								"type": "null"
							}
						],
						"title": "Description",
						"description": "What this recipe is for."
					},
					"spec": {
						"$ref": "#/components/schemas/GenerationSpec",
						"description": "The spec to save."
					}
				},
				"additionalProperties": false,
				"type": "object",
				"required": ["name", "spec"],
				"title": "RecipeCreate",
				"description": "A named, reusable spec saved for later."
			},
			"RecipeList": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/RecipeRead"
						},
						"type": "array",
						"title": "Items",
						"description": "Recipes in this page, newest first."
					},
					"total": {
						"type": "integer",
						"title": "Total",
						"description": "Total recipes matching the query."
					},
					"limit": {
						"type": "integer",
						"title": "Limit",
						"description": "Page size applied."
					},
					"offset": {
						"type": "integer",
						"title": "Offset",
						"description": "Offset applied."
					}
				},
				"additionalProperties": false,
				"type": "object",
				"required": ["total", "limit", "offset"],
				"title": "RecipeList",
				"description": "A page of recipes."
			},
			"RecipeRead": {
				"properties": {
					"id": {
						"type": "string",
						"title": "Id",
						"description": "Recipe identifier."
					},
					"name": {
						"type": "string",
						"title": "Name",
						"description": "Unique recipe name."
					},
					"description": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Description",
						"description": "What it is for."
					},
					"spec": {
						"$ref": "#/components/schemas/GenerationSpec",
						"description": "The stored spec."
					},
					"created_at": {
						"type": "string",
						"format": "date-time",
						"title": "Created At",
						"description": "When the recipe was saved."
					},
					"updated_at": {
						"type": "string",
						"format": "date-time",
						"title": "Updated At",
						"description": "When the recipe was last modified."
					}
				},
				"additionalProperties": false,
				"type": "object",
				"required": ["id", "name", "spec", "created_at", "updated_at"],
				"title": "RecipeRead",
				"description": "A saved recipe as returned by the API."
			},
			"RecipeRunResponse": {
				"properties": {
					"job_id": {
						"type": "string",
						"title": "Job Id",
						"description": "Identifier of the queued job."
					},
					"status": {
						"type": "string",
						"enum": ["queued", "processing", "complete", "complete_with_warnings", "failed"],
						"title": "Status",
						"description": "Job status at submission time."
					},
					"poll_url": {
						"type": "string",
						"title": "Poll Url",
						"description": "Path to poll for the job's result."
					}
				},
				"additionalProperties": false,
				"type": "object",
				"required": ["job_id", "status", "poll_url"],
				"title": "RecipeRunResponse",
				"description": "Acknowledgement for a job launched from a saved recipe."
			},
			"RecipeUpdate": {
				"properties": {
					"name": {
						"anyOf": [
							{
								"type": "string",
								"maxLength": 128,
								"minLength": 1
							},
							{
								"type": "null"
							}
						],
						"title": "Name",
						"description": "New name."
					},
					"description": {
						"anyOf": [
							{
								"type": "string",
								"maxLength": 512
							},
							{
								"type": "null"
							}
						],
						"title": "Description",
						"description": "New description."
					},
					"spec": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/GenerationSpec"
							},
							{
								"type": "null"
							}
						],
						"description": "New spec."
					}
				},
				"additionalProperties": false,
				"type": "object",
				"title": "RecipeUpdate",
				"description": "Partial update of a saved recipe.\n\nEvery field is optional, so a caller can rename a recipe without resending\nthe spec, or edit the spec without touching the name."
			},
			"SpectralAnalysis": {
				"properties": {
					"fundamental_detected_hz": {
						"type": "number",
						"exclusiveMinimum": 0.0,
						"title": "Fundamental Detected Hz",
						"description": "Fundamental frequency measured in the output, in Hz."
					},
					"ceiling_violation_db": {
						"type": "number",
						"title": "Ceiling Violation Db",
						"description": "Energy found above spectral_ceiling_hz, in dB relative to the ceiling. 0.0 or less means the roll-off held; positive values mean the ceiling was exceeded."
					},
					"attack_measured_ms": {
						"type": "number",
						"minimum": 0.0,
						"title": "Attack Measured Ms",
						"description": "Attack time measured in the output, in milliseconds."
					},
					"floor_slope_db_per_oct": {
						"type": "number",
						"maximum": 0.0,
						"title": "Floor Slope Db Per Oct",
						"description": "Measured high-pass slope below spectral_floor_hz, in dB per octave. Negative values mean energy falls as frequency drops: -6 is a single-pole filter, -24 is a 4-pole one. Values >= 0 mean no usable high-pass was applied."
					},
					"tilt_measured_db_per_octave": {
						"type": "number",
						"maximum": 24.0,
						"minimum": -24.0,
						"title": "Tilt Measured Db Per Octave",
						"description": "Measured spectral slope above the fundamental, in dB per octave. Compared against the spec's spectral_tilt_db_per_octave to confirm the requested darkening or brightening was actually applied."
					}
				},
				"additionalProperties": false,
				"type": "object",
				"required": [
					"fundamental_detected_hz",
					"ceiling_violation_db",
					"attack_measured_ms",
					"floor_slope_db_per_oct",
					"tilt_measured_db_per_octave"
				],
				"title": "SpectralAnalysis",
				"description": "Measurements taken from a generated file *after* post-processing.\n\nThese are the empirical values, not the requested ones. Comparing them\nagainst the originating :class:`GenerationSpec` tells you whether the hard\nconstraints were actually honoured.\n\nThe first three fields come from point measurements (a pitch estimate, a\nband energy ratio, an envelope segmentation). :attr:`floor_slope_db_per_oct`\nand :attr:`tilt_measured_db_per_octave` come from least-squares slope fits\nover the spectrum, which is why they are reported as dB per octave rather\nthan absolute dB."
			},
...
			"ValidationError": {
				"properties": {
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"type": "array",
						"title": "Location"
					},
					"msg": {
						"type": "string",
						"title": "Message"
					},
					"type": {
						"type": "string",
						"title": "Error Type"
					},
					"input": {
						"title": "Input"
					},
					"ctx": {
						"type": "object",
						"title": "Context"
					}
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			}
		}
	},
...
}